fix dep cycle for Rust
